Move prime sampling to a different file
This is going to make things simpler as the prime logic gets more
complicpated.
